+++ changelog.md
@@ -15,6 +15,7 @@
 - Fixed issue with `newton` iteration method when `attempt_best` was True and the routine would not exit if the error of the previous and current iterations were only equal
 - The `callback` for the `muller` iteration method will now correctly be passed the value of the evaluated function for the iteration, rather than the error.
 - Fixed description of `root_tol` and `attempt_best` in `iterate_to_root` docstring
+- Changes default `root_tol` to 0 for `secant`, `newton` and `muller` functions
 
 ### 1.1.11 - 22/Dec/2021
 - Fixed error when using `romb` integration method when supplied with a derivative function that returns a constant value